feat: adjusted server locale and jackson
