+++ src/main/java/com/luizmatias/findadev/api/dtos/responses/MatchDTO.java
@@ -1,0 +1,12 @@
+package com.luizmatias.findadev.api.dtos.responses;
+
+import java.util.Date;
+
+public record MatchDTO(
+        Long id,
+        UserDTO clientUser,
+        UserDTO developerUser,
+        Date matchedAt,
+        Date unmatchedAt
+) {
+}
