+++ src/main/resources/application.properties
@@ -1,6 +1,14 @@
+#Locale
+spring.web.locale=en_EN
+spring.web.locale-resolver=fixed
+
+#Database
 spring.jpa.hibernate.ddl-auto=update
 spring.datasource.url=jdbc:postgresql://localhost:5432/findadev
 spring.datasource.username=postgres
 spring.datasource.password=root
 spring.jpa.show-sql=true
+
+#Jackson
 spring.jackson.property-naming-strategy=SNAKE_CASE
+spring.jackson.default-property-inclusion=non_null
